requirement.txt
Flask
Flask-CORS
celery
redis
pandas
numpy
sentence-transformers
faiss-cpu
tqdm
openpyxl